feat: implement keycloak UserAccessToken and refactor keycloak package
Move token validation logic out into a separate function to enable
re-use and various other cleanups.
