+++ internal/keycloak/useraccesstoken.go
@@ -1,0 +1,52 @@
+package keycloak
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"path"
+	"time"
+
+	"github.com/google/uuid"
+	"go.opentelemetry.io/otel"
+	"golang.org/x/oauth2"
+)
+
+// UserAccessToken queries Keycloak given the user UUID, and returns an access
+// token. Authorized party for this token is auth-server. Authorization is done
+// by the Lagoon API.
+func (c *Client) UserAccessToken(ctx context.Context,
+	userUUID *uuid.UUID) (string, error) {
+	// set up tracing
+	ctx, span := otel.Tracer(pkgName).Start(ctx, "UserAccessToken")
+	defer span.End()
+	// get user token
+	tokenURL := *c.baseURL
+	tokenURL.Path = path.Join(tokenURL.Path,
+		`/auth/realms/lagoon/protocol/openid-connect/token`)
+	userConfig := oauth2.Config{
+		ClientID:     c.clientID,
+		ClientSecret: c.clientSecret,
+		Endpoint: oauth2.Endpoint{
+			TokenURL: tokenURL.String(),
+		},
+	}
+	ctx = context.WithValue(ctx, oauth2.HTTPClient, &http.Client{
+		Timeout: 10 * time.Second,
+	})
+	userToken, err := userConfig.Exchange(ctx, "",
+		// https://datatracker.ietf.org/doc/html/rfc8693#section-2.1
+		oauth2.SetAuthURLParam("grant_type",
+			"urn:ietf:params:oauth:grant-type:token-exchange"),
+		// https://www.keycloak.org/docs/latest/securing_apps/#_token-exchange
+		oauth2.SetAuthURLParam("requested_subject", userUUID.String()))
+	if err != nil {
+		return "", fmt.Errorf("couldn't get user token: %v", err)
+	}
+	// parse and extract verified attributes
+	_, err = c.validateTokenClaims(userToken)
+	if err != nil {
+		return "", fmt.Errorf("couldn't validate token claims: %v", err)
+	}
+	return userToken.AccessToken, nil
+}
